Replace Community page with Scout and Stats
