+++ front-end/src/components/BottomNav.jsx
@@ -1,12 +1,13 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
-import { Home, MessageCircle, Users, User } from 'lucide-react';
+import { Home, MessageCircle, Users, BarChart2, User } from 'lucide-react';
 
 export default function BottomNav({ clanIcon }) {
   const items = [
     { to: '/', label: 'Home', icon: Home },
     { to: '/chat', label: 'Chat', icon: MessageCircle },
-    { to: '/community', label: 'Community', icon: Users },
+    { to: '/scout', label: 'Scout', icon: Users },
+    { to: '/stats', label: 'Stats', icon: BarChart2 },
     { to: '/account', label: 'Account', icon: User },
   ];
