+++ front-end/src/pages/Scout.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Scout() {
+  return <div className="p-4">Coming soon...</div>;
+}
